refactor: clean up LogError
