+++ internal/utils/error.go
@@ -1,6 +1,7 @@
 package utils
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -11,6 +12,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 )
 
 type Error struct {
@@ -144,4 +146,43 @@ func SetAttributeErrorsFromServerResponse(
 			diags.AddAttributeError(attributePath, summary, errorDetail)
 		}
 	}
+}
+
+type ErrorResponse struct {
+	CorrelationId string              `json:"correlationId"`
+	ErrorCode     string              `json:"errorCode"`
+	ErrorMessage  string              `json:"errorMessage"`
+	ErrorDetails  map[string][]string `json:"errorDetails"`
+}
+
+// newErrorResponse generates a new ErrorResponse object from an api response.
+func newErrorResponse(body io.Reader) (*ErrorResponse, error) {
+	buf := new(strings.Builder)
+	_, err := io.Copy(buf, body)
+	if err != nil {
+		return nil, err
+	}
+
+	errorResponse := ErrorResponse{}
+
+	jsonErr := json.Unmarshal([]byte(buf.String()), &errorResponse)
+	if jsonErr != nil {
+		return nil, jsonErr
+	}
+
+	return &errorResponse, nil
+}
+
+// LogError prints the passed errorResponse as a Terraform error log.
+func LogError(ctx context.Context, httpResponse *http.Response, summary string) {
+	errorResponse, err := newErrorResponse(httpResponse.Body)
+	if err == nil {
+		tflog.Error(ctx, summary)
+	}
+
+	tflog.Error(
+		ctx,
+		summary,
+		map[string]interface{}{"ErrorResponse": errorResponse},
+	)
 }
